sandbox/async-react/pyproject.toml
[project]
name = "async-react"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
